+++ backend/common/settings/generator.go
@@ -705,6 +705,13 @@ func GenerateYaml() {
 	Config.Server.Sources = []*Source{
 		{
 			Path: ".",
+			Config: SourceConfig{
+				Conditionals: ConditionalFilter{
+					ItemRules: []ConditionalIndexConfig{
+						{}, // Add empty sample rule to show all fields in generated YAML
+					},
+				},
+			},
 		},
 	}