+++ frontend/tests/playwright/noauth/share.mobile.spec.ts
@@ -3,7 +3,7 @@ import { test, expect } from "../test-setup";
 
 test.use({viewport: { width: 750, height: 750 }}); // mobile viewport
 test("share download multiple files", async ({ page, checkForErrors, context }) => {
-  await page.goto("/files/files/");
+  await page.goto("/files/exclude/files/");
   await expect(page).toHaveTitle("Graham's Filebrowser - Files - playwright-files");
 
   const shareHash = await page.evaluate(() => localStorage.getItem('shareHash'));